Finalized V5 Contract Integration
